requirements.txt
asgiref==3.8.1
Django==5.1.7
django-extensions==3.2.3
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
